refactor : 패키지 구조 변경 1회
domain/user 패키지 하위에 유저 및 인증 관련 패키지들을 모두 이동시켰습니다. 인증관련 패키지는 이후에 이동될 수 있습니다.
