+++ src/main/java/CoCoNut_was/domains/user/controller/AuthController.java
@@ -1,4 +1,4 @@
-package CoCoNut_was.controller;
+package CoCoNut_was.domains.user.controller;
 
 /*
 로그인 및 로그아웃을 다루고 JWT 토큰 발행을 관리하는 컨트롤러
