+++ src/main/java/CoCoNut_was/domains/user/controller/UserController.java
@@ -1,13 +1,13 @@
-package CoCoNut_was.controller;
+package CoCoNut_was.domains.user.controller;
 
 /*
 회원가입(추가), 회원조회, 회원수정, 회원삭제를 다루는 컨트롤러
  */
 
-import CoCoNut_was.dto.UserReqDto;
-import CoCoNut_was.dto.UserResDto;
-import CoCoNut_was.entity.User;
-import CoCoNut_was.service.UserService;
+import CoCoNut_was.domains.user.dto.UserReqDto;
+import CoCoNut_was.domains.user.dto.UserResDto;
+import CoCoNut_was.domains.user.entity.User;
+import CoCoNut_was.domains.user.service.UserService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
