+++ src/main/java/CoCoNut_was/domains/user/dto/UserResDto.java
@@ -1,6 +1,6 @@
-package CoCoNut_was.dto;
+package CoCoNut_was.domains.user.dto;
 
-import CoCoNut_was.entity.User;
+import CoCoNut_was.domains.user.entity.User;
 import lombok.Builder;
 import lombok.Data;
