+++ src/main/java/CoCoNut_was/domains/user/entity/Role.java
@@ -1,4 +1,4 @@
-package CoCoNut_was.entity;
+package CoCoNut_was.domains.user.entity;
 
 public enum Role {
     ROLE_USER, ROLE_BUSINESS
