+++ src/main/java/CoCoNut_was/domains/user/service/AuthService.java
@@ -1,0 +1,4 @@
+package CoCoNut_was.domains.user.service;
+
+public class AuthService {
+}
