--- src/main/java/CoCoNut_was/service/AuthService.java
@@ -1,4 +1,0 @@
-package CoCoNut_was.service;
-
-public class AuthService {
-}
